refactor(composer_rpc): split trace.rs into directory module (4.5)
Split trace.rs (1474 LOC) into:
- types.rs (244 LOC): CallTraceNode, DetectedCall, ProxyInfo, ProxyLookup,
  ABI bindings, selector helpers
- walker.rs (343 LOC): walk_trace_tree, recurse_children
- proxy.rs (141 LOC): resolve_proxy_info, ephemeral proxy extraction
- mod.rs (805 LOC): re-exports + tests

Public API preserved via re-exports. No file mixes more than one
responsibility (plan §4 goal).

616/616 tests, clippy clean.
